+++ CHANGELOG.md
@@ -46,6 +46,7 @@ This changelog keeps track of work items that have been completed and are ready 
 ### Fixes
 
 - **General**: TODO ([#TODO](https://github.com/kedacore/http-add-on/issues/TODO))
+- **Interceptor**: Accept `TLS12`/`TLS13` TLS version format in addition to `1.2`/`1.3` for compatibility with KEDA and the operator ([#1718](https://github.com/kedacore/http-add-on/issues/1718))
 - **Scaler**: Fix scale-from-zero failing with 504 after long idle when using `requestRate` with large `window/granularity` ratios (>2000) ([#1692](https://github.com/kedacore/http-add-on/issues/1692))
 
 ### Deprecations